Add level "City" and experiment with images
We add a new level, "City", and put images in the answers. For that
change, we had to update inner HTML instead of inner text of the answer
buttons.

Since innerHTML canonicalizes the given input values, and we do not want
to keep re-rendering on every refresh, we had to use a dummy element to
canonicalize the answers' HTML ourselves. This might not be optimal, but
works for now.